+++ packages/tui/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added terminal background color query support for OSC 11 replies.
 
+### Fixed
+
+- Fixed WezTerm inline Kitty image rendering during full redraw fallbacks so image padding rows are reserved before the placement is drawn without regressing tall-image placement ([#5618](https://github.com/earendil-works/pi/issues/5618), [#4415](https://github.com/earendil-works/pi/issues/4415)).
+
 ## [0.79.3] - 2026-06-13
 
 ## [0.79.2] - 2026-06-12